feat(MainSidebar): renamed demoroutes and fixed exports. #464
